+++ library/src/schemas/nullable/nullableAsync.ts
@@ -13,15 +13,16 @@ export type NullableSchemaAsync<
   TWrapped extends BaseSchema | BaseSchemaAsync,
   TDefault extends
     | Input<TWrapped>
+    | null
     | undefined
-    | Promise<Input<TWrapped> | undefined> = undefined,
-  TOutput = Awaited<TDefault> extends undefined
+    | Promise<Input<TWrapped> | null | undefined> = undefined,
+  TOutput = Awaited<TDefault> extends undefined | null
     ? Output<TWrapped> | null
     : Output<TWrapped>
 > = BaseSchemaAsync<Input<TWrapped> | null, TOutput> & {
   schema: 'nullable';
   wrapped: TWrapped;
-  getDefault(): TDefault;
+  getDefault(): Promise<TDefault>;
 };
 
 /**
@@ -36,8 +37,9 @@ export function nullableAsync<
   TWrapped extends BaseSchema | BaseSchemaAsync,
   TDefault extends
     | Input<TWrapped>
+    | null
     | undefined
-    | Promise<Input<TWrapped> | undefined> = undefined
+    | Promise<Input<TWrapped> | null | undefined> = undefined
 >(
   wrapped: TWrapped,
   default_?: TDefault | (() => TDefault)
@@ -56,7 +58,7 @@ export function nullableAsync<
     /**
      * Returns the default value.
      */
-    getDefault() {
+    async getDefault() {
       return typeof default_ === 'function'
         ? (default_ as () => TDefault)()
         : (default_ as TDefault);